# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QA Dashboard API
  description: |
    API backend du QA Dashboard by Kimi 2.0.
    Gestion des métriques Testmo, synchronisation GitLab, exports, auth OAuth, et monitoring.
  version: 2.0.0
  contact:
    name: Neo-Logix QA Team
servers:
  - url: /api
    description: Serveur local (proxy Vite)
  - url: http://localhost:3001/api
    description: Serveur backend direct
...
security:
  - bearerAuth: []
  - adminToken: []
paths:
  /auth/gitlab:
    get:
      tags: [Auth]
      summary: Lancer l'authentification OAuth GitLab
      responses:
        302:
          description: Redirection vers GitLab OAuth
        501:
          description: OAuth non configuré
  /auth/gitlab/callback:
    get:
      tags: [Auth]
      summary: Callback OAuth GitLab
      parameters:
        - name: code
          in: query
          schema:
            type: string
      responses:
        302:
          description: Redirection vers le frontend avec token
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token d'accès
      responses:
        200:
          description: Nouveau token généré
        401:
          description: Refresh token invalide
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Profil utilisateur connecté
      security:
        - bearerAuth: []
      responses:
        200:
          description: Profil utilisateur
...
  /export/csv:
    post:
      tags: [Export]
      summary: Exporter les métriques en CSV
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: integer
                milestones:
                  type: object
      responses:
        200:
          description: Fichier CSV
  /export/excel:
    post:
      tags: [Export]
      summary: Exporter les métriques en Excel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: integer
                milestones:
                  type: object
      responses:
        200:
          description: Fichier Excel (.xlsx)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token